backend/requirements.txt
fastapi
uvicorn
textblob
pydantic
nrclex
langdetect
spacy
pandas
python-multipart
firebase-admin
google-generativeai
python-dotenv
